Projeto Ingresso

+++ Trabalho POO - Java 01/ProjetoIngresso/src/projetoingresso/IngressoVIP.java
@@ -10,7 +10,7 @@ public class IngressoVIP extends Ingresso {
     
 
     public IngressoVIP() {
-        this.valor = (float) (this.valor + 10.00);
+        this.valor = this.valor + 10.00f;
     }
 
     public float getValorVip() {
